Some build.zig improvements
Instruct Zigs ConfigHeader generator to use the existing config.h.in in
CMake style to avoid repetition of a lot of the variables.

Set a minimum set of variables to get this thing to build.

Basic support for building as static / shared lib, copying
BUILD_SHARED_LIBS option from CMake config.

Pull in libyang as a zig package dep (I wrote a build.zig for it
separately). Also pcre2, which has a build.zig from upstream repo but
was missing build.zig.zon, so using a fork of the repo for now where I
added that.

Building the sysrepoctl executable to verify that linking and everything
works correctly.

Fix some indent.